chore(map): relocate map logic to map.go
